refactor: relocate frontend tests to dedicated tests directory and update configuration
