+++ src/ravioli/frontend/vite.config.ts
@@ -7,10 +7,14 @@ export default defineConfig({
         target: 'http://localhost:8000',
         changeOrigin: true,
       }
+    },
+    fs: {
+      allow: ['../../..']
     }
   },
   test: {
     environment: 'happy-dom',
     globals: true,
+    include: ['../../../tests/ravioli/frontend/**/*.{test,spec}.ts']
   }
 })
