+++ tests/README.md
@@ -6,21 +6,11 @@ This directory mirrors the structure of `src/ravioli/` and contains all unit tes
 
 All commands should be run from the project root (`ravioli/`):
 
-```bash
-# Run all tests
+# Run backend tests
 uv run pytest tests/
 
-# Run with verbose output
-uv run pytest tests/ -v
-
-# Run a specific test file
-uv run pytest tests/ravioli/backend/core/test_dbt.py
-
-# Run a specific test function
-uv run pytest tests/ravioli/backend/core/test_dbt.py::test_run_dbt_command_success
-
-# Run with coverage report
-uv run pytest tests/ --cov=src/ravioli
+# Run frontend tests
+npm --prefix src/ravioli/frontend test
 ```
 
 ## Directory Structure
